+++ src/main/java/wordle/domain/generator/IndexGenerator.java
@@ -1,6 +1,5 @@
-package wordle.domain;
+package wordle.domain.generator;
 
 public interface IndexGenerator {
-
     int get(int size);
 }
