+++ src/main/java/wordle/domain/record/GameRecord.java
@@ -1,10 +1,11 @@
-package wordle.domain;
+package wordle.domain.record;
+
+import wordle.domain.Result;
 
 import java.util.Collections;
 import java.util.List;
 
 public class GameRecord {
-
     private final List<Result> results;
 
     public GameRecord(List<Result> results) {
