+++ src/main/java/wordle/domain/record/GameRecords.java
@@ -1,4 +1,4 @@
-package wordle.domain;
+package wordle.domain.record;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,9 +6,6 @@ import java.util.List;
 
 public class GameRecords {
     private final List<GameRecord> records = new ArrayList<>();
-
-    public GameRecords() {
-    }
 
     public void add(GameRecord gameRecord) {
         this.records.add(gameRecord);
