+++ src/main/java/wordle/util/InputReader.java
@@ -3,7 +3,6 @@ package wordle.util;
 import java.util.Scanner;
 
 public class InputReader {
-
     public String getUserInput() {
         Scanner sc = new Scanner(System.in);
         return sc.nextLine();
