--- src/test/java/wordle/GameMachineTest.java
@@ -1,5 +1,0 @@
-package wordle;
-
-class GameMachineTest {
-
-}
